requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
PyYAML==6.0.2
tenacity==9.0.0
